[antigravity] create a ping pong player struct. It should have a dependency on a tracker interface and a coinflip interface. it should have a single method called "play". it should have a constructor that takes the tracker, the coinflipper, and a name. When Play is called, it should register itself by name with the tracker, and ask if it's serving the ball. If it is serving, it should send the tracker a hit. if it is not serving, it should wait to receive from the tracker. when receiving a ball from the tracker, the player should perform a coin flip. If heads, it hits the ball back. if tails, it misses. If it hits, it should then wait & can either receive a ball back or be told the game is over. if it misses, it tells the tracker the game is over. In the game over case, it should stop. This should all happen asynchronously. Players should have a "Done" channel that is closed when the game is over.
yes, use imptest to generate the mocks and use those

please also generate the tests, mocks, and run the tests

use imptest.Start, and validate that the function ends as expected at the end of the test. Also run multiple players simultaneously in this test

this test will fail sometimes due to non-deterministic call order for the register function. In order to handle that, add a "Within(d duration)" method to generated imptest code, such that you can call `trackerImp.Within(time.Second).ExpectCallTo.Register("Ping").Resolve()`. If a different call is made before the expected call, instead of failing, push that call onto a threadsafe queue. Fresh calls to `trackerImp.Within().ExpectCallTo` or just `trackerImp.ExpectCallTo` should check that queue and use and remove matches if they find them, before trying to wait for new calls.

only remove and return the call if it matches both name and expected args

don't unlock until returning from GetCall. just put the unlock in a defer.

perform the timeout on the entire GetCall call, rather than on individual waits on callChan. It doesn't matter if we pull 1 or 100 items off of the callChan, we should timeout from GetCall after the specified duration.

the generated code locks and unlocks too often in `GetCall`. Stop generating those lines - just lock at the start of the function and unlock in the defer